third_party/blink/web_tests/http/tests/devtools/version-controller-expected.txt

Tests inspector version controller.


Running: testMethodsToRunToUpdateVersion
Testing methods to run to upgrade from 0 to 0.
Methods to run: []

Testing methods to run to upgrade from 0 to 1.
Methods to run: ["updateVersionFrom0To1"]

Testing methods to run to upgrade from 0 to 2.
Methods to run: ["updateVersionFrom0To1","updateVersionFrom1To2"]

Testing methods to run to upgrade from 0 to 3.
Methods to run: ["updateVersionFrom0To1","updateVersionFrom1To2","updateVersionFrom2To3"]

Testing methods to run to upgrade from 1 to 1.
Methods to run: []

Testing methods to run to upgrade from 1 to 2.
Methods to run: ["updateVersionFrom1To2"]

Testing methods to run to upgrade from 1 to 3.
Methods to run: ["updateVersionFrom1To2","updateVersionFrom2To3"]

Testing methods to run to upgrade from 2 to 2.
Methods to run: []

Testing methods to run to upgrade from 2 to 3.
Methods to run: ["updateVersionFrom2To3"]


Running: testClearBreakpointsWhenTooMany
Starting test with 0 breakpoints and 500 allowed at max.
Number of breakpoints left in the setting after the test: 0.

Starting test with 1 breakpoints and 500 allowed at max.
Number of breakpoints left in the setting after the test: 1.

Starting test with 2 breakpoints and 500 allowed at max.
Number of breakpoints left in the setting after the test: 2.

Starting test with 499 breakpoints and 500 allowed at max.
Number of breakpoints left in the setting after the test: 499.

Starting test with 500 breakpoints and 500 allowed at max.
Number of breakpoints left in the setting after the test: 500.

Starting test with 501 breakpoints and 500 allowed at max.
Number of breakpoints left in the setting after the test: 0.

Starting test with 1000 breakpoints and 500 allowed at max.
Number of breakpoints left in the setting after the test: 0.
